requirements.txt
# Core app
streamlit
python-dotenv
PyPDF2
tenacity

# LangChain stack
langchain
langchain-classic
langchain-community
langchain-google-genai
langchain-groq
langchain-text-splitters
sentence-transformers<3.0

# Vector store
faiss-cpu==1.9.0

# Google AI
google-generativeai

# Evaluation
ragas
datasets

# Utilities
numpy<2.0
pandas